notifiaction dynamique

+++ app/Http/Controllers/ProduitController.php
@@ -54,7 +54,7 @@ class ProduitController extends Controller
      */
     public function create()
     {
-        
+        $this->authorize('create', Produit::class);
         return view('produits.create');
     }
      
@@ -123,6 +123,7 @@ class ProduitController extends Controller
      */
     public function update(Request $request,Produit $produit)
     {
+        $this->authorize('update', $produit);
         $request->validate([
             'categorie' => 'required',
             'produit' => 'required',
@@ -142,6 +143,7 @@ class ProduitController extends Controller
      */
     public function destroy(Produit $produit)
     {
+        $this->authorize('delete', $produit);
         $produit->delete();
 
         return redirect()->route('produits.index')
